fix(examples) fix build
