+++ package.json
@@ -105,7 +105,7 @@
   "size-limit": [
     {
       "path": "cdn/agent.js",
-      "limit": "8.1 KB"
+      "limit": "8 KB"
     },
     {
       "path": "cdn/agent.d.js",
